// src/api/generated/model/postScbQrCreate200Data.ts
/**
 * Generated by orval v8.0.0-rc.2 🍺
 * Do not edit manually.
 * Smart City Hub
 * A comprehensive API
 * OpenAPI spec version: 1.0.0
 */
import type { PostScbQrCreate200DataStatus } from './postScbQrCreate200DataStatus';
import type { PostScbQrCreate200DataData } from './postScbQrCreate200DataData';

export type PostScbQrCreate200Data = {
  status: PostScbQrCreate200DataStatus;
  data: PostScbQrCreate200DataData;
};